refactor: 💡 move ignore null to another directory
