+++ lib/masking/config/target_columns/method/type/extension/ignore_null.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+module Masking
+  class Config
+    class TargetColumns
+      class Method
+        module Type
+          module Extension
+            module IgnoreNull
+              def call(sql_value)
+                if sql_value == 'NULL'
+                  sequence! if respond_to?(:sequence!, true)
+                  return 'NULL'
+                end
+
+                super
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+end
